temp

+++ utils.py
@@ -233,7 +233,7 @@ class TrainDataVer2(Dataset):
     def __getitem__(self, index):
         uid, bid = self.ub_pairs[index]
         prob_iids = np.array(self.ui_graph[uid].todense()).reshape(-1)
-        prob_iids_bundle = np.array(self.bi_graph[bid].todense()).reshape(-1)
+        prob_iids_bundle = np.array(self.bi_graph[bid].todense(), dtype=int).reshape(-1)
         return uid, prob_iids, prob_iids_bundle
 
     def __len__(self):
@@ -272,12 +272,12 @@ class TrainDataVer3(Dataset):
     def __getitem__(self, index):
         uid, bid = self.ub_pairs[index]
         prob_iids = np.array(self.ui_graph[uid].todense()).reshape(-1)
-        prob_iids_bundle = np.array(self.bi_graph[bid].todense()).reshape(-1)
+        prob_iids_bundle = np.array(self.bi_graph[bid].todense(), dtype=int).reshape(-1)
         while True:
             nbid = np.random.choice(self.num_bundle)
             if self.ub_graph[uid, nbid] == 0:
                 break
-        prob_iids_bundle -= np.array(self.bi_graph[nbid].todense()).reshape(-1)
+        prob_iids_bundle -= np.array(self.bi_graph[nbid].todense(), dtype=int).reshape(-1)
         return uid, prob_iids, prob_iids_bundle
 
     def __len__(self):
